+++ src/tkc/value.c
@@ -1403,3 +1403,228 @@ ret_t value_toggle_bit(value_t* v, value_t* result, uint32_t n) {
 
   return ret;
 }
+
+ret_t value_bit_not(value_t* v, value_t* result) {
+  ret_t ret = RET_OK;
+  return_value_if_fail(v != NULL && result != NULL, RET_BAD_PARAMS);
+
+  switch (v->type) {
+    case VALUE_TYPE_INT8: {
+      int8_t vv = value_int8(v);
+      value_set_int8(result, ~vv);
+      break;
+    }
+    case VALUE_TYPE_UINT8: {
+      uint8_t vv = value_uint8(v);
+      value_set_uint8(result, ~vv);
+      break;
+    }
+    case VALUE_TYPE_INT16: {
+      int16_t vv = value_int16(v);
+      value_set_int16(result, ~vv);
+      break;
+    }
+    case VALUE_TYPE_UINT16: {
+      uint16_t vv = value_uint16(v);
+      value_set_uint16(result, ~vv);
+      break;
+    }
+    case VALUE_TYPE_INT32: {
+      int32_t vv = value_int32(v);
+      value_set_int32(result, ~vv);
+      break;
+    }
+    case VALUE_TYPE_UINT32: {
+      uint32_t vv = value_uint32(v);
+      value_set_uint32(result, ~vv);
+      break;
+    }
+    case VALUE_TYPE_INT64: {
+      int64_t vv = value_int64(v);
+      value_set_int64(result, ~vv);
+      break;
+    }
+    case VALUE_TYPE_UINT64: {
+      uint64_t vv = value_uint64(v);
+      value_set_uint64(result, ~vv);
+      break;
+    }
+    default: {
+      ret = RET_BAD_PARAMS;
+      break;
+    }
+  }
+
+  return ret;
+}
+
+ret_t value_bit_or(value_t* v, value_t* other, value_t* result) {
+  ret_t ret = RET_OK;
+  uint32_t type = 0;
+  return_value_if_fail(result != NULL, RET_BAD_PARAMS);
+  return_value_if_fail(v != NULL && other != NULL, RET_BAD_PARAMS);
+
+  type = tk_max_int((int)(v->type), (int)(other->type));
+  switch (type) {
+    case VALUE_TYPE_INT8: {
+      int8_t vv = value_int8(v) | value_int8(other);
+      value_set_int8(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT8: {
+      uint8_t vv = value_uint8(v) | value_uint8(other);
+      value_set_uint8(result, vv);
+      break;
+    }
+    case VALUE_TYPE_INT16: {
+      int16_t vv = value_int16(v) | value_int16(other);
+      value_set_int16(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT16: {
+      uint16_t vv = value_uint16(v) | value_uint16(other);
+      value_set_uint16(result, vv);
+      break;
+    }
+    case VALUE_TYPE_INT32: {
+      int32_t vv = value_int32(v) | value_int32(other);
+      value_set_int32(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT32: {
+      uint32_t vv = value_uint32(v) | value_uint32(other);
+      value_set_uint32(result, vv);
+      break;
+    }
+    case VALUE_TYPE_INT64: {
+      int64_t vv = value_int64(v) | value_int64(other);
+      value_set_int64(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT64: {
+      uint64_t vv = value_uint64(v) | value_uint64(other);
+      value_set_uint64(result, vv);
+      break;
+    }
+    default: {
+      ret = RET_BAD_PARAMS;
+      break;
+    }
+  }
+
+  return ret;
+}
+
+ret_t value_bit_and(value_t* v, value_t* other, value_t* result) {
+  ret_t ret = RET_OK;
+  uint32_t type = 0;
+  return_value_if_fail(result != NULL, RET_BAD_PARAMS);
+  return_value_if_fail(v != NULL && other != NULL, RET_BAD_PARAMS);
+
+  type = tk_max_int((int)(v->type), (int)(other->type));
+  switch (type) {
+    case VALUE_TYPE_INT8: {
+      int8_t vv = value_int8(v) & value_int8(other);
+      value_set_int8(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT8: {
+      uint8_t vv = value_uint8(v) & value_uint8(other);
+      value_set_uint8(result, vv);
+      break;
+    }
+    case VALUE_TYPE_INT16: {
+      int16_t vv = value_int16(v) & value_int16(other);
+      value_set_int16(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT16: {
+      uint16_t vv = value_uint16(v) & value_uint16(other);
+      value_set_uint16(result, vv);
+      break;
+    }
+    case VALUE_TYPE_INT32: {
+      int32_t vv = value_int32(v) & value_int32(other);
+      value_set_int32(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT32: {
+      uint32_t vv = value_uint32(v) & value_uint32(other);
+      value_set_uint32(result, vv);
+      break;
+    }
+    case VALUE_TYPE_INT64: {
+      int64_t vv = value_int64(v) & value_int64(other);
+      value_set_int64(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT64: {
+      uint64_t vv = value_uint64(v) & value_uint64(other);
+      value_set_uint64(result, vv);
+      break;
+    }
+    default: {
+      ret = RET_BAD_PARAMS;
+      break;
+    }
+  }
+
+  return ret;
+}
+
+ret_t value_bit_nor(value_t* v, value_t* other, value_t* result) {
+  ret_t ret = RET_OK;
+  uint32_t type = 0;
+  return_value_if_fail(result != NULL, RET_BAD_PARAMS);
+  return_value_if_fail(v != NULL && other != NULL, RET_BAD_PARAMS);
+
+  type = tk_max_int((int)(v->type), (int)(other->type));
+  switch (type) {
+    case VALUE_TYPE_INT8: {
+      int8_t vv = value_int8(v) ^ value_int8(other);
+      value_set_int8(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT8: {
+      uint8_t vv = value_uint8(v) ^ value_uint8(other);
+      value_set_uint8(result, vv);
+      break;
+    }
+    case VALUE_TYPE_INT16: {
+      int16_t vv = value_int16(v) ^ value_int16(other);
+      value_set_int16(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT16: {
+      uint16_t vv = value_uint16(v) ^ value_uint16(other);
+      value_set_uint16(result, vv);
+      break;
+    }
+    case VALUE_TYPE_INT32: {
+      int32_t vv = value_int32(v) ^ value_int32(other);
+      value_set_int32(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT32: {
+      uint32_t vv = value_uint32(v) ^ value_uint32(other);
+      value_set_uint32(result, vv);
+      break;
+    }
+    case VALUE_TYPE_INT64: {
+      int64_t vv = value_int64(v) ^ value_int64(other);
+      value_set_int64(result, vv);
+      break;
+    }
+    case VALUE_TYPE_UINT64: {
+      uint64_t vv = value_uint64(v) ^ value_uint64(other);
+      value_set_uint64(result, vv);
+      break;
+    }
+    default: {
+      ret = RET_BAD_PARAMS;
+      break;
+    }
+  }
+
+  return ret;
+}