+++ docs/changes.md
@@ -1,6 +1,7 @@
 2022/09/26
   * 增加value\_lshift/value\_rshift/value\_lshift\_r/value\_rshift\_r等函数。
   * 增加value\_set\_bit/value\_get\_bit/value\_toggle\_bit等函数。
+  * 增加value\_bit\_not/value\_bit\_and/value\_biti\_or等函数。
 
 2022/09/23
   * 完善rbuffer\_read\_string接口(感谢雨欣提供补丁)